common: move FileHandler class in common module
